+++ internal/clients/kibana/alerting_test.go
@@ -1,12 +1,19 @@
 package kibana
 
 import (
+	"context"
+	"fmt"
+	"io"
+	"net/http"
+	"strings"
 	"testing"
 	"time"
 
 	"github.com/elastic/terraform-provider-elasticstack/generated/alerting"
 	"github.com/elastic/terraform-provider-elasticstack/internal/models"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/stretchr/testify/require"
+	gomock "go.uber.org/mock/gomock"
 )
 
 func Test_ruleResponseToModel(t *testing.T) {
@@ -119,6 +126,127 @@ func Test_ruleResponseToModel(t *testing.T) {
 			model := ruleResponseToModel(tt.spaceId, tt.ruleResponse)
 
 			require.Equal(t, tt.expectedModel, model)
+		})
+	}
+}
+
+func Test_CreateUpdateAlertingRule(t *testing.T) {
+	ctrl := gomock.NewController(t)
+
+	getApiClient := func() (ApiClient, *alerting.MockAlertingAPI) {
+		apiClient := NewMockApiClient(ctrl)
+		apiClient.EXPECT().SetAlertingAuthContext(gomock.Any()).DoAndReturn(func(ctx context.Context) context.Context { return ctx })
+		alertingClient := alerting.NewMockAlertingAPI(ctrl)
+		apiClient.EXPECT().GetAlertingClient().DoAndReturn(func() (alerting.AlertingAPI, error) { return alertingClient, nil })
+		return apiClient, alertingClient
+	}
+
+	tests := []struct {
+		name        string
+		testFunc    func(ctx context.Context, apiClient ApiClient, rule models.AlertingRule) (*models.AlertingRule, diag.Diagnostics)
+		client      ApiClient
+		rule        models.AlertingRule
+		expectedRes *models.AlertingRule
+		expectedErr string
+	}{
+		{
+			name:     "CreateAlertingRule should not crash when backend returns 4xx",
+			testFunc: CreateAlertingRule,
+			client: func() ApiClient {
+				apiClient, alertingClient := getApiClient()
+				alertingClient.EXPECT().CreateRule(gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(func(ctx context.Context, spaceId string, ruleId string) alerting.ApiCreateRuleRequest {
+					return alerting.ApiCreateRuleRequest{ApiService: alertingClient}
+				})
+				alertingClient.EXPECT().CreateRuleExecute(gomock.Any()).DoAndReturn(func(r alerting.ApiCreateRuleRequest) (*alerting.RuleResponseProperties, *http.Response, error) {
+					return nil, &http.Response{
+						StatusCode: 401,
+						Body:       io.NopCloser(strings.NewReader("some error")),
+					}, nil
+				})
+				return apiClient
+			}(),
+			rule:        models.AlertingRule{},
+			expectedRes: nil,
+			expectedErr: "some error",
+		},
+		{
+			name:     "UpdateAlertingRule should not crash when backend returns 4xx",
+			testFunc: UpdateAlertingRule,
+			client: func() ApiClient {
+				apiClient, alertingClient := getApiClient()
+				alertingClient.EXPECT().UpdateRule(gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(func(ctx context.Context, spaceId string, ruleId string) alerting.ApiUpdateRuleRequest {
+					return alerting.ApiUpdateRuleRequest{ApiService: alertingClient}
+				})
+				alertingClient.EXPECT().UpdateRuleExecute(gomock.Any()).DoAndReturn(func(r alerting.ApiUpdateRuleRequest) (*alerting.RuleResponseProperties, *http.Response, error) {
+					return nil, &http.Response{
+						StatusCode: 401,
+						Body:       io.NopCloser(strings.NewReader("some error")),
+					}, nil
+				})
+				return apiClient
+			}(),
+			rule:        models.AlertingRule{},
+			expectedRes: nil,
+			expectedErr: "some error",
+		},
+		{
+			name:     "CreateAlertingRule should not crash when backend returns an empty response and HTTP 200",
+			testFunc: CreateAlertingRule,
+			client: func() ApiClient {
+				apiClient, alertingClient := getApiClient()
+				alertingClient.EXPECT().CreateRule(gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(func(ctx context.Context, spaceId string, ruleId string) alerting.ApiCreateRuleRequest {
+					return alerting.ApiCreateRuleRequest{ApiService: alertingClient}
+				})
+				alertingClient.EXPECT().CreateRuleExecute(gomock.Any()).DoAndReturn(func(r alerting.ApiCreateRuleRequest) (*alerting.RuleResponseProperties, *http.Response, error) {
+					return nil, &http.Response{
+						StatusCode: 200,
+						Body:       io.NopCloser(strings.NewReader("everything seems fine")),
+					}, nil
+				})
+				return apiClient
+			}(),
+			rule:        models.AlertingRule{},
+			expectedRes: nil,
+			expectedErr: "empty response",
+		},
+		{
+			name:     "UpdateAlertingRule should not crash when backend returns an empty response and HTTP 200",
+			testFunc: UpdateAlertingRule,
+			client: func() ApiClient {
+				apiClient, alertingClient := getApiClient()
+				alertingClient.EXPECT().UpdateRule(gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(func(ctx context.Context, spaceId string, ruleId string) alerting.ApiUpdateRuleRequest {
+					return alerting.ApiUpdateRuleRequest{ApiService: alertingClient}
+				})
+				alertingClient.EXPECT().UpdateRuleExecute(gomock.Any()).DoAndReturn(func(r alerting.ApiUpdateRuleRequest) (*alerting.RuleResponseProperties, *http.Response, error) {
+					return nil, &http.Response{
+						StatusCode: 200,
+						Body:       io.NopCloser(strings.NewReader("everything seems fine")),
+					}, nil
+				})
+				return apiClient
+			}(),
+			rule:        models.AlertingRule{},
+			expectedRes: nil,
+			expectedErr: "empty response",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			rule, diags := tt.testFunc(context.Background(), tt.client, tt.rule)
+
+			if tt.expectedRes == nil {
+				require.Nil(t, rule)
+			} else {
+				require.Equal(t, tt.expectedRes, rule)
+			}
+
+			if tt.expectedErr != "" {
+				require.NotEmpty(t, diags)
+				if !strings.Contains(diags[0].Detail, tt.expectedErr) {
+					require.Fail(t, fmt.Sprintf("Diags ['%s'] should contain message ['%s']", diags[0].Detail, tt.expectedErr))
+				}
+			}
 		})
 	}
 }